### 2018-03-23 #641 @whedon
[PRE REVIEW]: CLIMLAB: a Python toolkit for interactive, process-oriented climate modeling
**Submitting author:** @brian-rose (<a href="http://orcid.org/0000-0002-9961-3821">Brian Rose</a>)
**Repository:** <a href="https://github.com/brian-rose/climlab/" target ="_blank">https://github.com/brian-rose/climlab/</a>
**Version:** v0.6.4
