### 2018-03-26 f49d3d6
Add affiliation to JOSS paper
