### 2018-03-26 #641 @brian-rose
I pushed an update to the climlab repo that fixes the missing affiliation issue. Not sure if I need to take some other action here?
