### 2018-03-26 2e6e212
Fix formatting and missing references in JOSS manuscript
