+++ paper.bib
@@ -48,3 +48,21 @@
 	Volume = {91},
 	Year = {2005},
 }
+
+@article{Emanuel:1991a,
+	Author = {Kerry A. Emanuel},
+	Journal = {J. Atmos. Sci.},
+	Pages = {2313--2314},
+	Title = {A Scheme for Representing Cumulus Convection in Large-Scale Models},
+	Volume = {48},
+	Year = {1991},
+}
+
+@article{Emanuel:1999a,
+	Author = {Kerry A. Emanuel and Marina {\v Z}ivkovi{\'c}-Rothman},
+	Journal = {J. Atmos. Sci.},
+	Pages = {1766--1782},
+	Title = {Development and Evaluation of a Convection Scheme for Use in Climate Models},
+	Volume = {56},
+	Year = {1999},
+}
