### 2018-03-26 #641 @brian-rose
Ok there were some missing refs and formatting issues. Just pushed a new update.
