### 2018-03-26 #641 @brian-rose
Better but not picking up two missing refs that I just added to `paper.bib`
