### 2018-03-26 #641 @lheagy
👋 Hi @Chilipp, @andreas-h, with your background in climate and atmospheric science as well as experience with python, your input on this JOSS submission would be much appreciated! Would you be willing to review this submission for JOSS?
