### 2018-04-08 #659 @lheagy
👋 Thanks @Chilipp and @dhhagan for being willing to review. There is a checklist above for each of you as you go through the review process. It works through general checks, installation, functionality and the quality of the paper. Please let me know if you have any questions or if I can be of assistance in the review process!
