+++ paper.bib
@@ -1,6 +1,6 @@
 
 @article{Jeevanjee:2017a,
-	Author = {Nadir Jeevanjee and Pedram Hassanzadeh and Spencer A. Hill and Aditi Sheshadri},
+	Author = {Nadir Jeevanjee and Pedram Hassanzadeh and Spencer Hill and Aditi Sheshadri},
 	Doi = {10.1002/2017MS001038},
 	Journal = {J. Adv. Model. Earth Syst.},
 	Pages = {1760--1771},
@@ -11,6 +11,7 @@
 
 @article{Held:2005fk,
 	Author = {Isaac M. Held},
+    Doi = {10.1175/BAMS-86-11-1609},
 	Journal = {Bull. Amer. Meteor. Soc.},
 	Number = {11},
 	Pages = {1609--1614},
@@ -42,7 +43,7 @@
 @article{Clough:2005a,
 	Author = {Clough, S.A. and M.W. Shepard and E.J. Mlawer and J.S. Delamere and M.J. Iacono and K. Cady-Pereira and S. Boukabara and P.D. Brown},
 	Doi = {10.1016/j.jqsrt.2004.05.058},
-	Journal = {J. Quant. Spectrosocopy Radiative Transfer},
+	Journal = {J. Quant. Spectroscopy Radiative Transfer},
 	Pages = {233--244},
 	Title = {Atmospheric radiative transfer modeling: a summary of the {AER} codes},
 	Volume = {91},
@@ -51,15 +52,17 @@
 
 @article{Emanuel:1991a,
 	Author = {Kerry A. Emanuel},
+    Doi = {10.1175/1520-0469(1991)048<2313:ASFRCC>2.0.CO;2},
 	Journal = {J. Atmos. Sci.},
 	Pages = {2313--2314},
 	Title = {A Scheme for Representing Cumulus Convection in Large-Scale Models},
 	Volume = {48},
 	Year = {1991},
-}
+    }
 
 @article{Emanuel:1999a,
 	Author = {Kerry A. Emanuel and Marina {\v Z}ivkovi{\'c}-Rothman},
+    Doi = {10.1175/1520-0469(1999)056<1766:DAEOAC>2.0.CO;2},
 	Journal = {J. Atmos. Sci.},
 	Pages = {1766--1782},
 	Title = {Development and Evaluation of a Convection Scheme for Use in Climate Models},
