### 2018-04-10 #659 @brian-rose
Thanks @Chilipp for the help with the references and the Python 3 update. I merged the PR.

Will add some community guidelines next.
