### 2018-04-11 brian-rose/climlab#68 @brian-rose
Collecting information about usage of climlab
If you are using climlab (for teaching, research, publications, outreach, or anything else!), we want to know about it!

I would like to start keeping a collection of links and references to use cases. This will be helpful information for my funders (NSF), but also could be a great resource for new users looking to see climlab in action.

Entirely voluntary, but feel free to add your own use cases as comments in this thread. I will leave the issue open indefinitely.
