### 2018-04-11 brian-rose/climlab#68 @brian-rose
For classroom content, I currently have two overlapping collections of Jupyter notebooks:

- Graduate-level lecture notes at https://github.com/brian-rose/ClimateModeling_courseware
- Undergraduate-level lecture notes at http://www.atmos.albany.edu/facstaff/brose/classes/ENV415_Spring2018/handouts/  (git repo for these notes is [here](https://github.com/brian-rose/env-415-2018)
