### 2018-04-11 brian-rose/climlab#68 @brian-rose
Publications that use climlab include:

- [Rose et al. (2018), Ice Caps and Ice Belts: The Effects of Obliquity on Ice−Albedo Feedback, Astrophys. J. 846, doi:10.3847/1538-4357/aa8306](http://iopscience.iop.org/article/10.3847/1538-4357/aa8306/meta)
- Rencurrel and Rose (2018), Exploring the climatic response to wide variations in ocean heat transport on an aquaplanet, J. Climate (under revision)
