### 2018-04-11 22bc894
New page in docs on how to contribute
