### 2018-04-11 #659 @brian-rose
The documentation now includes a new page (http://climlab.readthedocs.io/en/latest/contributing.html) with explicit guidelines for
- reporting bugs and issues
- seeking help and support
- contributing bug fixes, new features, and documentation

There is also a [new open github issue](https://github.com/brian-rose/climlab/issues/68) where I'm inviting people to self-report their usage of CLIMLAB. I hope to collect references and links and make these available as additional resources for new users.
