### 2018-04-11 #659 @Chilipp
Thanks @brian-rose! That is fine with me. However, apparently you wrote the CONTRIBUTING in markdown, which doesn't render well in your docs. I'd recommend to fix this.
