### 2018-04-13 brian-rose/climlab#69 @dhhagan
Add instructions on running unit tests?
Hi Brian,

I'm reviewing your paper for JOSS and thought it may be nice to add a quick few lines showing others how to run the unittests with pytest. I see they are in the `climlab/tests/` directory, but it took a bit of digging to find them! I could have also just missed where you listed them, and if so, close without issue. Thanks!
