### 2018-04-13 #659 @dhhagan
Hey @brian-rose I suggested adding more clear instructions on running the tests, otherwise this paper is good to go!
