### 2018-04-15 #659 @lheagy
Thanks @dhhagan and @Chilipp for a speedy and thorough review! @brian-rose: please keep us updated on your progress on brian-rose/climlab#69 - we can proceed with publishing once addressed
