### 2018-04-16 brian-rose/climlab#69 @brian-rose
Agreed that some better instructions are needed here. I'll hopefully get this done this week. Thanks.
