### 2018-04-18 a0d98f9
New docs on building and testing
