+++ docs/source/support.rst
@@ -1,10 +1,10 @@
 .. highlight:: rst
 
-Support
+Acknowledgement
 =======
 
 
-Development of ``climlab`` and associated documentation is partially supported
+Development of CLIMLAB and associated documentation is partially supported
 by the National Science Foundation under Grant Number AGS-1455071 to Brian Rose.
 
 Any opinions, findings, and conclusions or recommendations expressed in this
