### 2018-04-18 brian-rose/climlab#69 @brian-rose
There is now a new section in the docs on [Building and Testing CLIMLAB](http://climlab.readthedocs.io/en/latest/contributing.html#building-and-testing-climlab) with explicit instructions.
