### 2018-04-18 #659 @brian-rose
As noted in https://github.com/brian-rose/climlab/issues/69, there is now a new section in the docs on [Building and Testing CLIMLAB](http://climlab.readthedocs.io/en/latest/contributing.html#building-and-testing-climlab) with explicit instructions.
