### 2018-04-19 #659 @lheagy
Thanks @brian-rose, @dhhagan, would you mind taking one last look and closing brian-rose/climlab#69 if it has been addressed?
