### 2018-04-19 brian-rose/climlab#69 @brian-rose
Closing this as based on feedback from @dhhagan in https://github.com/openjournals/joss-reviews/issues/659.
