### 2018-04-20 cfbf9ea
Add climlab.utils.thermo to docs!
