+++ docs/source/api/climlab.utils.rst
@@ -7,4 +7,5 @@ climlab.utils
    climlab.utils.constants
    climlab.utils.heat_capacity
    climlab.utils.legendre
+   climlab.utils.thermo
    climlab.utils.walk
