+++ conda-recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "climlab" %}
-{% set version = "0.6.5.dev0" %}
+{% set version = "0.6.5.dev1" %}
 
 package:
   name: {{ name|lower }}
@@ -9,7 +9,7 @@ source:
   path: ../
 
 build:
-  number: 5
+  number: 0
 
 requirements:
   build:
