### 2018-04-20 ddc6ea8
Increment version to 0.6.5
