+++ README.rst
@@ -194,7 +194,8 @@ Version 0.6.3 (released February 2018) brings under-the-hood improvements to the
 
 Version 0.6.4 (released February 2018) brings some bug fixes and a new ``climlab.couple()`` method to simplify creating complete models from components.
 
-Version 0.6.5 (released April 2018) adds some improved documentation, associated with the JOSS paper.
+Version 0.6.5 (released April 2018) adds some improved documentation, associated with publication of a meta-description paper in JOSS.
+
 
 The documentation_ was first created by Moritz Kreuzer (Potsdam Institut for Climate Impact Research) as part of a thesis project in Spring 2016.
 
@@ -226,7 +227,6 @@ See the accompanying LICENSE file.
    :target: http://climlab.readthedocs.io/en/latest/intro.html
    :alt: Documentation Status
 
-.. |JOSS| [status](http://joss.theoj.org/papers/6831d26a8ae32c11b8991bb848d0d4e2/status.svg)](http://joss.theoj.org/papers/6831d26a8ae32c11b8991bb848d0d4e2)
 =======
 
 
