+++ conda-recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "climlab" %}
-{% set version = "0.6.5.dev1" %}
+{% set version = "0.6.5" %}
 
 package:
   name: {{ name|lower }}
