### 2018-04-20 #659 @brian-rose
I just released climlab v0.6.5 including the updated documentation. It is archived on Zenodo under doi:10.5281/zenodo.1226360 ([direct link](https://zenodo.org/record/1226360#.Wto4sS_MxUM))
